transcribe_gcp/requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
google-cloud-speech==2.9.0
gunicorn==20.1.0
pydub==0.25.1